requirements.txt
numpy>=1.26.0
scikit-learn>=1.4.0
xgboost>=2.0.0
umap-learn>=0.5.6
hdbscan>=0.8.33
qdrant-client>=1.9.0
openai>=1.30.0
sentence-transformers>=2.7.0
langchain>=0.2.0
langgraph>=0.2.0
apscheduler>=3.10.0
optuna>=3.6.0
pandas>=2.2.0
matplotlib>=3.8.0
seaborn>=0.13.0
